requirements.txt
onnxruntime
mss
pydirectinput
pyqt5
requests
numpy
opencv-python
keyboard
pyyaml
pywin32
yt-dlp
